#1855 move functions to static Bond methods
